Added branch with my personal changes
This is a redesign of the internal data structure to make it more efficient.
Also added proper handling and issuing of events so that other mods can interact with it.  Specifically, this allows the Dispatcher mod to keep a correct list of train stops, and issue trains to stringy stops that rename themselves to the Dispatcher naming convention.